Move plugins/standalone.py to _internal/
