--- docs/api/plugins/standalone.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.plugins.standalone`
--------------------------------------
-
-.. automodule:: certbot.plugins.standalone
-   :members:
